+++ lib/simple_3d_renderer.dart
@@ -1,5 +1,6 @@
 export 'src/sp3d_camera.dart';
 export 'src/sp3d_faceobj.dart';
+export 'src/sp3d_gesture_detector.dart';
 export 'src/sp3d_light.dart';
 export 'src/sp3d_paint_image.dart';
 export 'src/sp3d_renderer.dart';